refactor: 예약 시간 관련 코드 구조 개선
- 예외 문구 상수로 처리
- 관리자와 사용자로 dto 세부 패키지 분리
- final 필드 설정 및 널 값 검증 추가
