+++ src/main/java/roomescape/reservationTime/dto/user/AvailableReservationTimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.reservationTime.dto;
+package roomescape.reservationTime.dto.user;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDate;
