+++ src/main/java/roomescape/reservationTime/dto/user/AvailableReservationTimeResponse.java
@@ -1,0 +1,11 @@
+package roomescape.reservationTime.dto.user;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import java.time.LocalTime;
+
+public record AvailableReservationTimeResponse(
+        Long timeId,
+        @JsonFormat(pattern = "HH:mm") LocalTime startAt,
+        Boolean alreadyBooked
+) {
+}
