+++ src/test/java/roomescape/reservationTime/dto/user/AvailableReservationTimeRequestTest.java
@@ -1,4 +1,4 @@
-package roomescape.reservationTime.dto;
+package roomescape.reservationTime.dto.user;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
